+++ vitest.config.ts
@@ -12,6 +12,8 @@ export default defineConfig({
         "/packages/generators/json/src/index.ts",
       "@schema-transformation-toolkit/generator-csv":
         "/packages/generators/csv/src/index.ts",
+      "@schema-transformation-toolkit/generator-toml":
+        "/packages/generators/toml/src/index.ts",
       "@schema-transformation-toolkit/generator-openapi":
         "/packages/generators/openapi/src/index.ts",
       "@schema-transformation-toolkit/generator-typescript":
@@ -24,6 +26,8 @@ export default defineConfig({
         "/packages/parsers/json/src/index.ts",
       "@schema-transformation-toolkit/parser-csv":
         "/packages/parsers/csv/src/index.ts",
+      "@schema-transformation-toolkit/parser-toml":
+        "/packages/parsers/toml/src/index.ts",
       "@schema-transformation-toolkit/parser-json-schema":
         "/packages/parsers/json-schema/src/index.ts",
       "@schema-transformation-toolkit/parser-typescript":